Create index.js
To edit the Township Small Business Webapp, you can open and modify the src/pages/index.js file. This is the main entry point for your Gatsby application and where you can make changes to the content of your homepage. Here's a sample of how you can edit this file: